Move generated rpc assets
